requirements.txt
streamlit
pdfplumber
pandas
google-generativeai
openpyxl
pytesseract
Pillow